% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{thesis_epub}
\alias{thesis_epub}
\title{Gera a tese em formato epub (ebook)}
\usage{
thesis_epub(...)
}
\arguments{
\item{...}{Argumentos adicionais passados para
\code{\link[bookdown]{epub_book}}.}
}
\value{
Um ebook em formato epub
}
\description{
Função usada no campo \code{output} do cabeçalho YAML do arquivo Rmd
principal para gerar uma versão ebook (epub) da tese.
}
\examples{
\dontrun{
# No cabeçalho YAML:
# output:
#   ufprdown::thesis_epub: default
}
}
