% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{thesis_gitbook}
\alias{thesis_gitbook}
\title{Gera a tese como página web (gitbook)}
\usage{
thesis_gitbook(...)
}
\arguments{
\item{...}{Argumentos adicionais passados para
\code{\link[bookdown]{gitbook}}.}
}
\value{
Uma página web gitbook
}
\description{
Função usada no campo \code{output} do cabeçalho YAML do arquivo Rmd
principal para gerar uma versão web interativa da tese.
}
\examples{
\dontrun{
# No cabeçalho YAML:
# output:
#   ufprdown::thesis_gitbook: default
}
}
