% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{thesis_word}
\alias{thesis_word}
\title{Gera a tese em formato Word}
\usage{
thesis_word(...)
}
\arguments{
\item{...}{Argumentos adicionais passados para
\code{\link[bookdown]{word_document2}}.}
}
\value{
Um documento Word
}
\description{
Função usada no campo \code{output} do cabeçalho YAML do arquivo Rmd
principal para gerar uma versão Microsoft Word da tese.
}
\examples{
\dontrun{
# No cabeçalho YAML:
# output:
#   ufprdown::thesis_word: default
}
}
